Update npm command in Dockerfile
The Dockerfile has been updated to use 'npm ci' instead of 'npm install'. This change ensures that the build is more reliable and faster, as it will only install the exact version specified in the package-lock.json, and it is suitable for automated environments such as test platforms, continuous integration, and deployment.